Fix fragment view, refactor string, added Scan and DetailScan Activity
