--- app/src/main/java/com/dicoding/aquaculture/view/ScanActivity.kt
@@ -1,177 +1,0 @@
-package com.dicoding.aquaculture.view
-
-import android.Manifest
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.net.Uri
-import android.os.Bundle
-import android.util.Log
-import android.view.MenuItem
-import android.view.View
-import android.widget.Toast
-import androidx.activity.enableEdgeToEdge
-import androidx.activity.result.PickVisualMediaRequest
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.activity.viewModels
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import com.dicoding.aquaculture.R
-import com.dicoding.aquaculture.data.utils.getImageUri
-import com.dicoding.aquaculture.data.utils.reduceFileImage
-import com.dicoding.aquaculture.data.utils.uriToFile
-import com.dicoding.aquaculture.databinding.ActivityScanBinding
-
-class ScanActivity : AppCompatActivity() {
-//    private val viewModel by viewModels<UploadStoryViewModel> {
-//        ViewModelFactory.getInstance(this)
-//    }
-    private lateinit var binding: ActivityScanBinding
-
-    private var currentImageUri: Uri? = null
-
-    private val requestPermissionLauncher =
-        registerForActivityResult(
-            ActivityResultContracts.RequestPermission()
-        ) { isGranted: Boolean ->
-            if (isGranted) {
-                Toast.makeText(this, "Permission request granted", Toast.LENGTH_LONG).show()
-            } else {
-                Toast.makeText(this, "Permission request denied", Toast.LENGTH_LONG).show()
-            }
-        }
-    private fun allPermissionsGranted() =
-        ContextCompat.checkSelfPermission(
-            this,
-            REQUIRED_PERMISSION
-        ) == PackageManager.PERMISSION_GRANTED
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityScanBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-//        supportActionBar?.title = getString(R.string.upload_story)
-
-        supportActionBar?.setDisplayHomeAsUpEnabled(true)
-
-        if (savedInstanceState != null) {
-            currentImageUri = savedInstanceState.getParcelable(KEY_IMAGE_URI)
-            currentImageUri?.let { showImage() }
-        }
-
-        if (!allPermissionsGranted()) {
-            requestPermissionLauncher.launch(REQUIRED_PERMISSION)
-        }
-
-        binding.galleryButton.setOnClickListener { startGallery() }
-        binding.cameraButton.setOnClickListener { startCamera() }
-        binding.uploadButton.setOnClickListener {
-            val uri = Uri.parse(currentImageUri.toString())
-            val intent = Intent(this, ScanDetailsActivity::class.java)
-            intent.putExtra(ScanDetailsActivity.EXTRA_IMAGE_URI, uri.toString())
-            startActivity(intent)
-        }
-
-//        viewModel.isLoading.observe(this, Observer { isLoading ->
-//            showLoading(isLoading)
-//        })
-//        viewModel.uploadResult.observe(this, Observer { message ->
-//            showToast(message)
-//            if (message == getString(R.string.upload_complete)) {
-//                navigateToMainActivity()
-//            }
-//        })
-    }
-
-    private fun startGallery() {
-        launcherGallery.launch(PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly))
-    }
-
-    private fun startCamera() {
-        currentImageUri = getImageUri(this)
-        launcherIntentCamera.launch(currentImageUri!!)
-    }
-
-    private val launcherIntentCamera = registerForActivityResult(
-        ActivityResultContracts.TakePicture()
-    ) { isSuccess ->
-        if (isSuccess) {
-            showImage()
-        }
-    }
-
-//    private fun uploadImage() {
-//        currentImageUri?.let { uri ->
-//            val imageFile = uriToFile(uri, this).reduceFileImage()
-//            Log.d("Image File", "showImage: ${imageFile.path}")
-//            val description = binding.descriptionEditText.text.toString().trim()
-//
-//            val requestBody = description.toRequestBody("text/plain".toMediaType())
-//            val requestImageFile = imageFile.asRequestBody("image/jpeg".toMediaType())
-//            val multipartBody = MultipartBody.Part.createFormData(
-//                "photo",
-//                imageFile.name,
-//                requestImageFile
-//            )
-//
-//            viewModel.upload(multipartBody, requestBody)
-//        } ?: showToast(getString(R.string.empty_image_warning))
-//    }
-
-    private val launcherGallery = registerForActivityResult(
-        ActivityResultContracts.PickVisualMedia()
-    ) { uri: Uri? ->
-        if (uri != null) {
-            currentImageUri = uri
-            showImage()
-        } else {
-            Log.d("Photo Picker", "No media selected")
-        }
-    }
-
-    private fun showImage() {
-        currentImageUri?.let {
-            Log.d("Image URI", "showImage: $it")
-            binding.previewImageView.setImageURI(it)
-        }
-    }
-
-    private fun showLoading(isLoading: Boolean) {
-        binding.progressIndicator.visibility = if (isLoading) View.VISIBLE else View.GONE
-    }
-    private fun showToast(message: String) {
-        Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
-    }
-    private fun navigateToDetailScanActivity() {
-        val intent = Intent(this, MainActivity::class.java)
-        intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-        startActivity(intent)
-        finish()
-    }
-
-    companion object {
-        private const val REQUIRED_PERMISSION = Manifest.permission.CAMERA
-        private const val KEY_IMAGE_URI = "KEY_IMAGE_URI"
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        if (item.itemId == android.R.id.home) {
-            onBackPressed()
-            return true
-        }
-        return super.onOptionsItemSelected(item)
-    }
-
-    override fun onSaveInstanceState(outState: Bundle) {
-        super.onSaveInstanceState(outState)
-        outState.putParcelable(KEY_IMAGE_URI, currentImageUri)
-    }
-
-    override fun onRestoreInstanceState(savedInstanceState: Bundle) {
-        super.onRestoreInstanceState(savedInstanceState)
-        currentImageUri = savedInstanceState.getParcelable(KEY_IMAGE_URI)
-        currentImageUri?.let { showImage() }
-    }
-}
