+++ app/src/main/java/com/dicoding/aquaculture/view/main/ScanFragment.kt
@@ -1,19 +1,49 @@
 package com.dicoding.aquaculture.view.main
 
+import android.Manifest
+import android.content.Intent
+import android.content.pm.PackageManager
+import android.net.Uri
 import android.os.Bundle
 import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.dicoding.aquaculture.R
-import com.dicoding.aquaculture.databinding.FragmentHomeBinding
+import android.widget.Toast
+import androidx.activity.result.PickVisualMediaRequest
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.core.content.ContextCompat
+import com.dicoding.aquaculture.data.utils.getImageUri
 import com.dicoding.aquaculture.databinding.FragmentScanBinding
+import com.dicoding.aquaculture.view.scan.ScanDetailsActivity
 
 
 class ScanFragment : Fragment() {
+//    private val viewModel by viewModels<UploadStoryViewModel> {
+//        ViewModelFactory.getInstance(this)
+//    }
     private var _binding: FragmentScanBinding? = null
     private val binding get() = _binding!!
+
+    private var currentImageUri: Uri? = null
+
+    private val requestPermissionLauncher =
+        registerForActivityResult(
+            ActivityResultContracts.RequestPermission()
+        ) { isGranted: Boolean ->
+            if (isGranted) {
+                Toast.makeText(requireContext(), "Permission request granted", Toast.LENGTH_LONG).show()
+            } else {
+                Toast.makeText(requireContext(), "Permission request denied", Toast.LENGTH_LONG).show()
+            }
+        }
+
+    private fun allPermissionsGranted() =
+        ContextCompat.checkSelfPermission(
+            requireContext(),
+            REQUIRED_PERMISSION
+        ) == PackageManager.PERMISSION_GRANTED
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -21,8 +51,111 @@ class ScanFragment : Fragment() {
     ): View? {
 
         _binding = FragmentScanBinding.inflate(inflater, container, false)
+
+        // If there is a saved instance, restore the image URI
+        savedInstanceState?.let {
+            currentImageUri = it.getParcelable(KEY_IMAGE_URI)
+            currentImageUri?.let { showImage() }
+        }
+
+        if (!allPermissionsGranted()) {
+            requestPermissionLauncher.launch(REQUIRED_PERMISSION)
+        }
+
+        binding.galleryButton.setOnClickListener { startGallery() }
+        binding.cameraButton.setOnClickListener { startCamera() }
+        binding.uploadButton.setOnClickListener {
+            val uri = Uri.parse(currentImageUri.toString())
+            val intent = Intent(requireContext(), ScanDetailsActivity::class.java)
+            intent.putExtra(ScanDetailsActivity.EXTRA_IMAGE_URI, uri.toString())
+            startActivity(intent)
+        }
+
+//        viewModel.isLoading.observe(this, Observer { isLoading ->
+//            showLoading(isLoading)
+//        })
+//        viewModel.uploadResult.observe(this, Observer { message ->
+//            showToast(message)
+//            if (message == getString(R.string.upload_complete)) {
+//                navigateToMainActivity()
+//            }
+//        })
+
         return binding.root
 
+    }
+
+
+
+    private fun startGallery() {
+        launcherGallery.launch(PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly))
+    }
+
+    private fun startCamera() {
+        currentImageUri = getImageUri(requireContext())
+        launcherIntentCamera.launch(currentImageUri!!)
+    }
+
+    private val launcherIntentCamera = registerForActivityResult(
+        ActivityResultContracts.TakePicture()
+    ) { isSuccess ->
+        if (isSuccess) {
+            showImage()
+        }
+    }
+
+    //    private fun uploadImage() {
+//        currentImageUri?.let { uri ->
+//            val imageFile = uriToFile(uri, this).reduceFileImage()
+//            Log.d("Image File", "showImage: ${imageFile.path}")
+//            val description = binding.descriptionEditText.text.toString().trim()
+//
+//            val requestBody = description.toRequestBody("text/plain".toMediaType())
+//            val requestImageFile = imageFile.asRequestBody("image/jpeg".toMediaType())
+//            val multipartBody = MultipartBody.Part.createFormData(
+//                "photo",
+//                imageFile.name,
+//                requestImageFile
+//            )
+//
+//            viewModel.upload(multipartBody, requestBody)
+//        } ?: showToast(getString(R.string.empty_image_warning))
+//    }
+
+    private val launcherGallery = registerForActivityResult(
+        ActivityResultContracts.PickVisualMedia()
+    ) { uri: Uri? ->
+        if (uri != null) {
+            currentImageUri = uri
+            showImage()
+        } else {
+            Log.d("Photo Picker", "No media selected")
+        }
+    }
+
+    private fun showImage() {
+        currentImageUri?.let {
+            Log.d("Image URI", "showImage: $it")
+            binding.previewImageView.setImageURI(it)
+        }
+    }
+
+    private fun showLoading(isLoading: Boolean) {
+        binding.progressIndicator.visibility = if (isLoading) View.VISIBLE else View.GONE
+    }
+
+    private fun showToast(message: String) {
+        Toast.makeText(requireContext(), message, Toast.LENGTH_SHORT).show()
+    }
+
+    companion object {
+        private const val REQUIRED_PERMISSION = Manifest.permission.CAMERA
+        private const val KEY_IMAGE_URI = "KEY_IMAGE_URI"
+    }
+
+    override fun onSaveInstanceState(outState: Bundle) {
+        super.onSaveInstanceState(outState)
+        outState.putParcelable(KEY_IMAGE_URI, currentImageUri)
     }
 
     override fun onDestroyView() {
