+++ app/src/main/java/com/dicoding/aquaculture/view/scan/ScanDetailsActivity.kt
@@ -1,14 +1,9 @@
-package com.dicoding.aquaculture.view
+package com.dicoding.aquaculture.view.scan
 
 import android.net.Uri
 import android.os.Bundle
 import android.util.Log
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import com.dicoding.aquaculture.R
-import com.dicoding.aquaculture.databinding.ActivityScanBinding
 import com.dicoding.aquaculture.databinding.ActivityScanDetailsBinding
 
 class ScanDetailsActivity : AppCompatActivity() {
