+++ app/src/main/res/layout/activity_scan_details.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.ScanDetailsActivity">
+    tools:context=".view.scan.ScanDetailsActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
