+++ app/src/main/res/layout/activity_signup.xml
@@ -20,8 +20,8 @@
             android:id="@+id/titleDaftar"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:text="Daftar"
-            android:textSize="36dp"
+            android:text="@string/daftar"
+            android:textSize="36sp"
             android:textColor="@color/black"
             android:textStyle="bold"
             android:layout_marginTop="16dp"
@@ -37,8 +37,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="@id/guidelineHorz2"
-            android:src="@drawable/logo"
-            />
+            android:src="@drawable/logo" />
 
         <TextView
             android:id="@+id/titleTextSignup"
@@ -55,8 +54,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="@id/guidelineHorz"
-            tools:alpha="100"
-            />
+            tools:alpha="100" />
 
         <TextView
             android:id="@+id/nameTextSignup"
@@ -65,13 +63,12 @@
             android:layout_marginStart="32dp"
             android:layout_marginTop="8dp"
             android:alpha="0"
-            android:text="Nama"
+            android:text="@string/name"
             android:textColor="@color/black"
             android:textSize="14sp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/titleTextSignup"
-            tools:alpha="100"
-            />
+            tools:alpha="100" />
 
         <com.google.android.material.textfield.TextInputLayout
             android:id="@+id/nameEditTextSignup"
@@ -84,16 +81,14 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@id/nameTextSignup"
             app:startIconDrawable="@drawable/ic_person"
-            tools:alpha="100"
-            >
+            tools:alpha="100">
 
                 <com.google.android.material.textfield.TextInputEditText
                     android:id="@+id/nameEditText"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:ems="10"
-                    android:inputType="text"
-                    />
+                    android:inputType="text" />
 
         </com.google.android.material.textfield.TextInputLayout>
 
@@ -104,13 +99,12 @@
             android:layout_marginStart="32dp"
             android:layout_marginTop="16dp"
             android:alpha="0"
-            android:text="Email"
+            android:text="@string/email"
             android:textColor="@color/black"
             android:textSize="14sp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/nameEditTextSignup"
-            tools:alpha="100"
-            />
+            tools:alpha="100" />
 
         <com.google.android.material.textfield.TextInputLayout
             android:id="@+id/emailEditTextSignup"
@@ -123,16 +117,14 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@id/emailTextSignup"
             app:startIconDrawable="@drawable/ic_email"
-            tools:alpha="100"
-            >
+            tools:alpha="100">
 
                 <com.dicoding.aquaculture.view.costumview.EmailEditText
                     android:id="@+id/emailEditText"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:ems="10"
-                    android:inputType="textEmailAddress"
-                    />
+                    android:inputType="textEmailAddress" />
 
         </com.google.android.material.textfield.TextInputLayout>
 
@@ -143,13 +135,12 @@
             android:layout_marginStart="32dp"
             android:layout_marginTop="16dp"
             android:alpha="0"
-            android:text="Password"
+            android:text="@string/password"
             android:textColor="@color/black"
             android:textSize="14sp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/emailEditTextSignup"
-            tools:alpha="100"
-            />
+            tools:alpha="100" />
 
         <com.google.android.material.textfield.TextInputLayout
             android:id="@+id/passwordEditTextSignup"
@@ -163,16 +154,14 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@id/passwordTextSignup"
             app:startIconDrawable="@drawable/ic_lock"
-            tools:alpha="100"
-            >
+            tools:alpha="100">
 
                 <com.dicoding.aquaculture.view.costumview.PasswordEditText
                     android:id="@+id/passwordEditText"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:ems="10"
-                    android:inputType="textPassword"
-                    />
+                    android:inputType="textPassword" />
 
         </com.google.android.material.textfield.TextInputLayout>
 
@@ -185,7 +174,8 @@
             android:layout_marginTop="32dp"
             android:enabled="false"
             android:alpha="0"
-            android:text="DAFTAR"
+            android:textAllCaps="true"
+            android:text="@string/daftar"
             android:textColor="@color/white"
             android:textSize="14sp"
             app:cornerRadius="6dp"
@@ -193,8 +183,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@id/passwordEditTextSignup"
-            tools:alpha="100"
-            />
+            tools:alpha="100" />
 
         <ProgressBar
             android:id="@+id/progressBar"
@@ -205,8 +194,7 @@
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
-            android:visibility="gone"
-            />
+            android:visibility="gone" />
 
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/guidelineHorz"
@@ -215,15 +203,14 @@
             android:orientation="horizontal"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintGuide_percent="0.3"
-            />
+            app:layout_constraintGuide_percent="0.3" />
+
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/guidelineHorz2"
             android:layout_width="wrap_content"
             android:layout_height="64dp"
             android:orientation="horizontal"
-            app:layout_constraintGuide_percent="0.15"
-            />
+            app:layout_constraintGuide_percent="0.15" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
