+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -3,8 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/mobile_navigation"
-    app:startDestination="@id/navigation_home"
-    >
+    app:startDestination="@id/navigation_home">
 
     <fragment
         android:id="@+id/navigation_home"
